Added option to ignore Content-Type header.
A kwarg parameter named ignore_content_type was added to get_pac and download_pac. The rationale for adding a new parameter is that it maintain the best backwards compatibility and ease of use. Also considered using accepting a specific value for allowed_content_types. However, I believe that this may be unexpected for current users. The arg was placed before the 'session' arg in order to maintain related args together. However, it risks breaking usages where all the parameters are positional.